Alternative solution to the Playlist.java project
Greetings! This is not much of a fix really, mostly I wanted to experiment with github and how to contribute to a project :)
I just changed the order of indices in lines 37,38 because for me it is easier to have it like this: indexA<indexB
and then got the file with the "get" method instead of writing the whole name. 
Sincerely, a student from CodeCademy :)